hid cursor

+++ tetris.py
@@ -1,5 +1,6 @@
-import sys, os, time, random, signal
+import sys, os, time, random, signal, curses
 from tetromino import *
+curses.initscr()
 win_rows, win_cols = os.popen('stty size', 'r').read().split()
 n = int(win_rows)
 m = int(win_cols)
@@ -32,10 +33,11 @@ def sig_handler(signum, frame):
 	if signum == signal.SIGINT:
 		print(clr + hom, end = '')
 		sys.stdout.flush()
+		curses.curs_set(1)
+		curses.endwin()
 		sys.exit()
 
 def main():
-
 	signal.signal(signal.SIGINT, sig_handler)
 	# clear screen
 	#print(clr + hom, end = '')
@@ -44,23 +46,35 @@ def main():
 	#	sys.stdout.flush()
 	grid = Grid(n,m)
 	minos = []
+	mino = None
 	t = -1	
-	fr = 0.001
+	fr = 0.000
+	curses.curs_set(0)
+	mino_oldpos = None
+	goal_col = 0
 	while(True):
 		t += 1
 		time.sleep(fr)
 		
-		for mino in minos:
+		if mino == None:
+			mino = Tetromino(random.choice('LJSZTOI'), grid, random.randint(1,4))
+			goal_col = random.randint(1,int(win_cols))
+		else:
 			mino.drop()	
-			if random.random() < 0.5:
+			if mino.get_pos()[0] < goal_col:
 				mino.right()
+			elif mino.get_pos()[0] > goal_col:
+				mino.left()
+			
+			if mino_oldpos == mino.get_pos():
+				mino = None
+			else:
+				mino_oldpos = mino.get_pos()
 			#if random.random() < 0.5:
 			#	if random.random() < 0.5:
 			#		mino.rotate('c')
 			#	else:
 			#		mino.rotate('cc')
-		if (t+1) % 10 == 0:
-			minos.append(Tetromino(random.choice('LJSZTOI'), grid, random.randint(1,4)))
 
 		grid.draw()
